requirements.txt

gunicorn
south
redis
django-celery
rarfile
MySQL-python
pillow
